DedicatedServerService: honor result code of relayServer.Start()
The method can return false under several circumstances, but the relay server
would just return as if the socket has been successfully created.

The MasterServer would believe, that the socket exists resulting in issues that
can be painful to debug. Reasons for a failure here can be sockets that are
used by other applications or no properly closed connections from a previous
attempt.

It might be even better to try a different port instead, but that should at least
give some information about, if this issue happens at all.